pyproject.toml
[tool.poetry]
name = "kargo"
version = "0.1.0"
description = ""
authors = ["hagunhee <hatchi10@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "kargo"}]


[tool.poetry.dependencies]
python = "^3.9.10"
django = "^4.1.6"
black = "^23.1.0"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.14.0"
requests = "^2.28.2"
hashids = "^1.3.1"

[tool.poetry.group.dev.dependencies]
black = {version = "^23.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
